==> analyse_alt/2 - Bearbeitung der Daten aus dem Live-XML-Copy1.ipynb <==
import pandas as pd
import numpy as np
# XML-Lib laden
import xml.etree.ElementTree as ET

## XML-Datei laden - Ergebnisse auf Gemeindeebene

root = ET.parse("./Rohdaten/wahl_gemeinden.xml").getroot()

columns = ["WK-Nr", "AGS", "Wahlberechtigte", "Wähler", "ungültige_1", "ungültige_2", 
           "gültige_1","gültige_2", "CDU_1", "DIE LINKE_1", "SPD_1", "AfD_1", 
          "GRÜNE_1", "NPD_1", "FDP_1", "FREIE WÄHLER_1", "Tierschutzpartei_1", "PIRATEN_1", "Die PARTEI_1",
           "BüSo_1", "ADPM_1", "Blaue #TeamPetry_1", "KPD_1", "ÖDP_1", "Die Humanisten_1", "PDV_1", 
           "Gesundheitsforschung_1", 
          "CDU_2", "DIE LINKE_2", "SPD_2", "AfD_2", 
          "GRÜNE_2", "NPD_2", "FDP_2", "FREIE WÄHLER_2", "Tierschutzpartei_2", "PIRATEN_2", "Die PARTEI_2",
           "BüSo_2", "ADPM_2", "Blaue #TeamPetry_2", "KPD_2", "ÖDP_2", "Die Humanisten_2", "PDV_2", 
           "Gesundheitsforschung_2"]

# finales Dataframe
df_ergebnisse = pd.DataFrame(columns=columns)

for elem in root: 
    # jetzt müssen sämtliche Variablen deklariert werden, damit sie durch if-else-Klauseln gefüllt werden können
    kreis = elem.attrib.get('KreisNr')
    gemeinde = elem.attrib.get('GemNr')
    wahlberechtigte = ""
    wähler = ""
    ungültige_1 = ""
    ungültige_2 = ""
    gültige_1 = ""
    gültige_2 = ""
    CDU_1 = "" 
    LINKE_1 =""
    SPD_1= "" 
    AfD_1 = ""
    GRÜNE_1 = ""
    NPD_1 = ""
    FDP_1 = ""
    FREIE_1 = ""
    Tierschutzpartei_1 = ""
    PIRATEN_1 = ""
    PARTEI_1 = "" 
    BüSo_1 = ""
    ADPM_1 = ""
    Blaue_1 = ""
    KPD_1 = "" 
    ÖDP_1 = ""
    Humanisten_1 = ""
    PDV_1 = ""
    Gesundheitsforschung_1 = ""
    
    #Zweitstimmen
    CDU_2 = "" 
    LINKE_2 =""
    SPD_2= "" 
    AfD_2 = ""
    GRÜNE_2 = ""
    NPD_2 = ""
    FDP_2 = ""
    FREIE_2 = ""
    Tierschutzpartei_2 = ""
    PIRATEN_2 = ""
    PARTEI_2 = "" 
    BüSo_2 = ""
    ADPM_2 = ""
    Blaue_2 = ""
    KPD_2 = "" 
    ÖDP_2 = ""
    Humanisten_2 = ""
    PDV_2 = ""
    Gesundheitsforschung_2 = ""
    
    
    
    
    
    
    for sub in elem: 
        # Identifikation der jeweiligen Information
        position = sub.attrib.get('Position')
        
        
        for subsub in sub:
            
            
           
            if position == "WB":
                wahlberechtigte = subsub.attrib.get("Stimmen")
            elif position == "WAE":
                wähler = subsub.attrib.get("Stimmen")
            elif position == "UNG":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    ungültige_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    ungültige_2 = subsub.attrib.get("Stimmen")
            elif position == "GUE":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    gültige_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    gültige_2 = subsub.attrib.get("Stimmen")
            elif position == "1":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    CDU_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    CDU_2 = subsub.attrib.get("Stimmen")
            elif position == "2":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    LINKE_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    LINKE_2 = subsub.attrib.get("Stimmen")
            elif position == "3":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    SPD_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    SPD_2 = subsub.attrib.get("Stimmen")
            elif position == "4":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    AfD_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    AfD_2 = subsub.attrib.get("Stimmen")
            elif position == "5":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    GRÜNE_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    GRÜNE_2 = subsub.attrib.get("Stimmen")
            elif position == "6":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    NPD_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    NPD_2 = subsub.attrib.get("Stimmen")
            elif position == "7":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    FDP_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    FDP_2 = subsub.attrib.get("Stimmen")
            elif position == "8":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    FREIE_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    FREIE_2 = subsub.attrib.get("Stimmen")
            elif position == "9":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    Tierschutzpartei_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    Tierschutzpartei_2 = subsub.attrib.get("Stimmen")
            elif position == "10":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    PIRATEN_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    PIRATEN_2 = subsub.attrib.get("Stimmen")
            elif position == "11":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    PARTEI_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    PARTEI_2 = subsub.attrib.get("Stimmen")
            elif position == "12":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    BüSo_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    BüSo_2 = subsub.attrib.get("Stimmen")
            elif position == "13":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    ADPM_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    ADPM_2 = subsub.attrib.get("Stimmen")
            elif position == "14":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    Blaue_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    Blaue_2 = subsub.attrib.get("Stimmen")
            elif position == "15":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    KPD_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    KPD_2 = subsub.attrib.get("Stimmen")
            elif position == "16":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    ÖDP_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    ÖDP_2 = subsub.attrib.get("Stimmen")
            elif position == "17":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    Humanisten_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    Humanisten_2 = subsub.attrib.get("Stimmen")
            elif position == "18":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    PDV_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    PDV_2 = subsub.attrib.get("Stimmen")
            elif position == "19":
                stimmart = subsub.attrib.get("Stimmart")
                if stimmart == "DIREKT": 
                    Gesundheitsforschung_1 = subsub.attrib.get("Stimmen")
                if stimmart == "LISTE": 
                    Gesundheitsforschung_2 = subsub.attrib.get("Stimmen")
                    
                    
                    
    df_ergebnisse = df_ergebnisse.append(pd.Series([
        kreis, gemeinde, wahlberechtigte, wähler, ungültige_1, ungültige_2, gültige_1, gültige_2, 
        CDU_1, LINKE_1, SPD_1, AfD_1, GRÜNE_1, NPD_1, FDP_1, FREIE_1, Tierschutzpartei_1, PIRATEN_1, 
        PARTEI_1, BüSo_1, ADPM_1, Blaue_1, KPD_1, ÖDP_1, Humanisten_1, PDV_1, Gesundheitsforschung_1, 
        CDU_2, LINKE_2, SPD_2, AfD_2, GRÜNE_2, NPD_2, FDP_2, FREIE_2, Tierschutzpartei_2, PIRATEN_2, 
        PARTEI_2, BüSo_2, ADPM_2, Blaue_2, KPD_2, ÖDP_2, Humanisten_2, PDV_2, Gesundheitsforschung_2         
    ], index=columns), ignore_index=True)

# alle Ergebnisse in int konvertieren
cols=[i for i in df_ergebnisse.columns]
for col in cols:
    df_ergebnisse[col]=pd.to_numeric(df_ergebnisse[col])
    df_ergebnisse[col]= df_ergebnisse[col].fillna(0).astype(int)

df_ergebnisse.tail()

### Alle Ergebnisse der Großstädte in einer AGS zusammenfassen

# WK in Großstädten einmal zusammenfassen
df_leipzig = df_ergebnisse[(df_ergebnisse["AGS"]>147130000) & (df_ergebnisse["AGS"]<147130010)]
df_dresden = df_ergebnisse[(df_ergebnisse["AGS"]>146120000) & (df_ergebnisse["AGS"]<146120010)]
df_zwickau = df_ergebnisse[(df_ergebnisse["AGS"]>145243300) & (df_ergebnisse["AGS"]<145243303)]
df_chemnitz = df_ergebnisse[(df_ergebnisse["AGS"]>145110000) & (df_ergebnisse["AGS"]<145110005)]

# AGS vergeben, über die später gruppiert wird
df_leipzig["AGS"] = 14713000
df_dresden["AGS"] = 14612000
df_zwickau["AGS"] = 14524330
df_chemnitz["AGS"] = 14511000

group_le = df_leipzig.groupby("AGS", as_index=False).sum()
group_dd = df_dresden.groupby("AGS", as_index=False).sum()
group_zw = df_zwickau.groupby("AGS", as_index=False).sum()
group_ch = df_chemnitz.groupby("AGS", as_index=False).sum()

df_ergebnisse = pd.concat([df_ergebnisse, group_le, group_dd, group_zw, group_ch], ignore_index=True, sort=False )

# 'Ergebnisse ausspielen
df_ergebnisse.to_csv("./Daten_clean/ergebnisse_lt2019.csv", sep=";", encoding="utf-8")

df_ergebnisse["WK-Nr"].unique()

test = df_ergebnisse[df_ergebnisse["AGS"]==14524170]
test
