==> analyse_alt/3 - Erstellung Testdatensatz LW2019.ipynb <==
import pandas as pd
import numpy as np

## Creating Dataframe with fake data
Dafür nehmen wir die Ergebnisse der letzten Landtagswahl und lassen das ganze zufällig schwanken. 
Das ist zwar teilweise verfälschend, weil wir dort andere Gebietszuschnitte hatten... 
... aber das müssen wir jetzt mal so hinnehmen. 





#testdaten_wahl = pd.read_csv("./Rohdaten/Testdaten_aus_LTW_2014.csv", encoding="utf-8", sep=";")
testdaten_wahl = pd.read_csv("./Daten_clean/ergebnisse_lt2019.csv", encoding="utf-8", sep=";")

testdaten_wahl.info()

### Merge mit den historischen Wahldaten

hist_df = pd.read_csv("./Daten_clean/historische_Wahldaten_bereinigt.csv", sep=";", encoding="utf-8")

tt = hist_df[hist_df["AGS"]==14524170]
tt

test_wahl_aus_lw2014_und_historischen_daten = testdaten_wahl.merge(hist_df, on="AGS", how="left")
test_wahl_aus_lw2014_und_historischen_daten.drop(columns=["Unnamed: 0_x"], inplace=True)

test_wahl_aus_lw2014_und_historischen_daten.columns


### Richtige Datentypen vergeben - alle Zahlen in np.int64

# alle Ergebnisse in int konvertieren
cols=[i for i in test_wahl_aus_lw2014_und_historischen_daten.columns if i!="WK-Name_LW2014"]
for col in cols:
   
    test_wahl_aus_lw2014_und_historischen_daten[col]=pd.to_numeric(test_wahl_aus_lw2014_und_historischen_daten[col])
    test_wahl_aus_lw2014_und_historischen_daten[col]= test_wahl_aus_lw2014_und_historischen_daten[col].fillna(0).astype(int)

test_wahl_aus_lw2014_und_historischen_daten

# Ausspielen
test_wahl_aus_lw2014_und_historischen_daten.to_csv("./Daten_clean/ergebnisse_lt_gesamt.csv", 
                                                  encoding="utf-8", sep=";")

# alte Version, vielleicht mal aufbewahren


# cols = ['WK-Nr', 'AGS', 'Wahlberechtigte', 'Wähler', 'gültige_1',
#        'CDU_1', 'DIE LINKE_1', 'SPD_1', 'FDP_1', 'GRÜNE_1', 'NPD_1',
#        'PIRATEN_1', 'BüSo_1', 'DSU_1', 'AfD_1', 'FREIE WÄHLER_1',
#        'Die PARTEI_1', 'Freie Bürger Leipzig_1', 'Konrad Skatula_1',
#        'Schaar_1', 'WAGNER_1', 'WV MFU_1', 'gültige_2', 'CDU_2', 'DIE LINKE_2',
#        'SPD_2', 'FDP_2', 'GRÜNE_2', 'NPD_2', 'Tierschutzpartei_2', 'PIRATEN_2',
#        'BüSo_2', 'DSU_2', 'AfD_2', 'pro Deutschland_2', 'FREIE WÄHLER_2',
#        'Die PARTEI_2', 'Unnamed: 0_y', 'Wahlberechtigte_EW2019',
#        'Wähler_EW2019', 'gültige_EW2019', 'CDU_EW2019', 'DIE LINKE_EW2019',
#        'SPD_EW2019', 'AfD_EW2019', 'GRÜNE_EW2019', 'FDP_EW2019',
#        'FREIE WÄHLER_EW2019', 'Wahlberechtigte_BW2017', 'Wähler_BW2017',
#        'gültige_2_BW2017', 'CDU_2_BW2017', 'DIE LINKE_2_BW2017',
#        'SPD_2_BW2017', 'AfD_2_BW2017', 'GRÜNE_2_BW2017', 'FDP_2_BW2017',
#        'FREIE WÄHLER_2_BW2017', 'WK-Nr_LW2014', 
#        'Wahlberechtigte_LW2014', 'Wähler_LW2014', 'gültige_1_LW2014',
#        'CDU_1_LW2014', 'DIE LINKE_1_LW2014', 'SPD_1_LW2014', 'FDP_1_LW2014',
#        'GRÜNE_1_LW2014', 'NPD_1_LW2014', 'PIRATEN_1_LW2014', 'BüSo_1_LW2014',
#        'DSU_1_LW2014', 'AfD_1_LW2014', 'FREIE WÄHLER_1_LW2014',
#        'Die PARTEI_1_LW2014', 'Freie Bürger Leipzig_1_LW2014',
#        'Konrad Skatula_1_LW2014', 'Schaar_1_LW2014', 'WAGNER_1_LW2014',
#        'WV MFU_1_LW2014', 'gültige_2_LW2014', 'CDU_2_LW2014',
#        'DIE LINKE_2_LW2014', 'SPD_2_LW2014', 'FDP_2_LW2014', 'GRÜNE_2_LW2014',
#        'NPD_2_LW2014', 'Tierschutzpartei_2_LW2014', 'PIRATEN_2_LW2014',
#        'BüSo_2_LW2014', 'DSU_2_LW2014', 'AfD_2_LW2014',
#        'pro Deutschland_2_LW2014', 'FREIE WÄHLER_2_LW2014',
#        'Die PARTEI_2_LW2014']

# test_wahl_aus_lw2014_und_historischen_daten[cols] = test_wahl_aus_lw2014_und_historischen_daten[cols].fillna(0).astype(np.int64)
