==> analyse_alt/4 Direktkandidaten.ipynb <==
import pandas as pd
import numpy as np
import json

df = pd.read_csv(r"C:\Users\Felix\Desktop\wahlkarte_react\datenanalyse\Daten_clean\Direktkandidaten_2019_Landtagswahl_Sachsen.csv", \
                   sep=";", encoding="utf-8")



data = df.copy()

data.head(10)

# Wahlkreis-Column von float nach int wandeln
data["Wahlkreis"] = data["Wahlkreis"].astype(int)

#Umanwandlung in JSON

# Schritt 1 - Tabelle wide machen
df_pivot = data.pivot(index="Partei", columns="Wahlkreis", values="Kandidat")

# Schritt 2 - Pivot-Tabelle in JSON konvertieren - ascii-Befehl erhält die Umlaute
df_pivot.to_json(r"C:\Users\Felix\Desktop\wahlkarte_react\datenanalyse\Daten_clean\Direktkandidaten_ZwischenschrittJSON.json", 
                      force_ascii=False)

        


# sorgt dafür, dass auch Special-Characters (ü,ä etc) mitgenommen werden. 

with open(r"C:\Users\Felix\Desktop\wahlkarte_react\datenanalyse\Daten_clean\Direktkandidaten_ZwischenschrittJSON.json", 'w', encoding='utf-8') as file:
    df_pivot.to_json(file, force_ascii=False)

df_pivot
